make abandoned test work in root crate
